refactor(files): Create controllers file
